Merge pull request #86 from gregarndt/ci_work
Bug 1160320 - Run containers in privileged mode

This allows docker-worker to start containers in a privileged mode to allow them privileges such as running docker-in-docker which helps for running the docker worker tests.

Both the worker and the task need to be configured to run as privileged for this to happen and tasks will be failed if they request privileged mode but the worker is not configured to allow it.

This starts the work of adding a "capabilities" object to the task payload that will tasks to define certain options that need to be available/enabled for the container. Later this will be expanded to include things for loopback devices, possible memory limits, etc.