Bug 1220738: Create endpoint for credentials update. r=pmoore
docker-worker will use task temporary credentials for taskcluster
requests on task behalf.

During task reclaim, the temporary credentials might be updated, and
if using taskcluster-proxy to forward requests, we should update the
credentials in the proxy as well.

We create a "/Credentials" endpoint which receive PUT requests with
a json body containing the new credentials data for update.