Regexp tweaks for slugid library
This tightens the regexp checks for slugs and uuids. Furthermore,
we now compile the regexp expressions upfront, to avoid compiling
with every decode/encode call.

Lastly, adjusted the tests for generating slugs, to generate 1000
slugs rather than one, to increase confidence in results. This
added 4ms to the test time, which seems a reasonable cost/benefit.