Print status to stderr. r=dustin
Within the "Do one job, and do it well" Unix philosophy, it is common to
glue commands together through shell pipes to achieve the desired
result. This implies commands must only print to stdout the result of
the processing of the program input, and any statuses or side effects
should go to stderr.

For taskcluster-cli, one example is the "api" subcommand, that wraps up
the taskcluster API and often prints a JSON string as the result of an
endpoint call. If we print other statuses to stdout, the JSON string
just becomes invalid and we cannot pipeline it to tools like jq [1].

[1] https://stedolan.github.io/jq/