Remove race conditions from stream piping
In several parts of the code there is a common pattern to combine nodejs
pipe method and waitForEvent:

stream1.pipe(stream2);
await waitForEvent(stream1, 'end');

It turns out this code is racy and may end up in a deadlock. If the 'end'
event arrives before the 'end' event, waitForEvent will get stuck
forever.

We solve it by using the promisepipe package, which supplies a promised
version of the pipe function and replace the code above with
promisepipe.