Include git revision SHA in version information
The git revision cannot be stored in the source, since you cannot
include the commit SHA in the source without affecting the commit
SHA itself. Therefore we include this information at link time
during the build. The git revision is then available via the
generic-worker --version command, and also in the
workerTypeMetadata that is published at the start of every task
log. A test has been added to make sure a SHA is available at
runtime. Sentry crash reports also include the revision SHA.