Fix upload retry tests
The strategy to simulate a network failure when testing the artifact
upload retry feature is to add a rule to iptables when we detect the
worker is uploading an artifact to S3.

Due to the asynchronous nature of nodejs, other parts of the worker may
run while waiting for I/O response, and with iptables blocking outgoing
traffic, the unrelated network calls will be blocked.

The solution is to block only the outgoing packets to Amazon S3 service.
We do so by downloading the IP ranges from Amazon and filtering those
that belong to the us-west-2 region and whose service is S3. We then
block those IP ranges.