Generate roles based on auth0 management API profile (#71)
* Generate roles based on auth0 management API profile

* Remove ldap and mozillian clients and replace with auth0

For client IDs, the `user_id` property of the auth0 management API is used since it's unique.
For logins with github, the nickname is appended to the `user_id`
(e.g., `github|0000|helfi92`) to give some meaning to the ID.

* Add credentials-response.yml back

* Add + and | to client ID schema

Reason: Make it the same as the regex in the auth service
https://github.com/taskcluster/taskcluster-auth/commit/57e7da04fe0
a49cba8347fd44e4927f36ac4c05f#diff-0f8cdc9cea2017cf8d085950183ff441

* Update tests

* Enhance scanner

* Take into consideration a client ID with many slashes

* Fix github test

* Add encode util to limit encoding

* Refactor to be cleaner

- New handler methods
- Remove usage of mozilla-auth0 in the scanner file
- Use the limiting encoding utility

* Fix tests

* Update and add new tests

* Remove % from client ID schema if we stop using slashes

github IDs will now be of the form github|0000|helfi92
I previously added '%' because I was under the assumption that github
IDs would be of the form github|0000/helfi92 which would necessitate to
encode the forward slash.

* Add better names to variables and refactor code

* Change encoding and decoding format and add tests

The change was made to follow the format outlined in:
https://github.com/taskcluster/taskcluster-auth/pull/137#issuecomment-367841083

* Refactor converting profile to identity for testing, and add tests

* add some (non-passing) tests for encoding/decoding clientIds

* Add error handling for handler and scanner

* Fix some tests

* verify that all legal clientId characters except / are unencoded, and all illegal characters are not

* Add bang symbol to schema

* Enhance the encoding function