Add scopes and selected scope view (#68)
* Add clients and client view

* Have the scope match function be a utility

* Add constants and prop-types

* Give DataTable the power to render a list of strings

Prior it was only rendering a list of objects.

* Create ClientScopesTable and RoleScopesTable components

Each of these components has the abiltiy to render
a list of scopes or a list of roles/clients. I created
these 2 in 1 components in order to use the same memoize function
as well as to reduce duplication of the code since a rather big
portion of the component code was going to be the same.

* Add the route /auth/scopes and create Scopes and Scope view

Also a small nit to the Scope view

* Remove DateChooser component

This component got here by accident when rebasing

* Skip putting tabs in the URL

so we can have clean URLs.