Remove rootUrl and switch to relative paths for schema id (#39)
[BREAKING] use SchemaSet ast the top-level concept

This changes the interface to the library such that users first construct a SchemaSet, then create a validator from that SchemaSet.  In the process, we define "abstract" and "absolute" schemas -- the former with $ids that are not URLs.

This change also requires that uses of `$ref` in schemas be relative, not absolute, and not start with `/` either.

Changes:
* Remove rootUrl and switch to relative paths for schema id
* README update
* rename taskclusterSchemas to abstractSchemas
* revert to just rendering constants
* factor all of the publishing nonsense out to a single file
* check that references are relative and don't start with /
* only construct the Ajv object when making a validator
* rename validate.js to schemaset.js
* validate an aboslute schema
* fix invalid_test to actually test things
* check that error messages have absolute URL in them
..and move SchemaSet-specific tests to a dedicated test file