Bug 1463260 - [BREAKING] Make references and schemas redeployability-friendly (#100)
This library now requires taskcluster-lib-validate 11.0.0 or higher, and thus requires a schemaset instead of a validator as input.

The `input` and `output` values of each API method must now be simple filenames (`foo-request.yaml`) without any path; they are suffixed to the schema URL, service name, and version at runtime.

The references generated by this library are now entirely relative to the rootUrl -- that is, they need not be specialized per deployment of Taskcluster.  Specifically, the `input` and `output` properties in `api-

Changes:
* Bug 1463260 - Fix rootUrl for docs writing
* make schema.{input,output} always relative URIs
relative to the service schema, <rootUrl>/schemas/<serviceName>
* upgrade taskcluster-lib-validate