Remove use of baseUrl
This removes the concept of a baseUrl, requiring a rootUrl instead.  The
taskcluster-lib-urls library is used to create URLs for API endpoints,
based on rootUrl, service name, api version, and method path.

The serviceName and apiVersion are considered non-overridable properties
of the geneated class, while rootUrl is a mandatory constructor
argument.

API references are also updated in this commit, to ensure that
everything has a serviceName defined.

This is a breaking change, so warrants a major version bump.