Bug 1515140 - support building from a monorepo
This is a pretty awful hack, but is the minimal change to get us running
from a monorepo, with lots of room to follow up after that.  The worst
sin is, it builds docker images containing dependencies for all
services, but only running one service.  This adds about 200MB to each
docker images, which is awful but something we can do while
transitioning.