[UI] Remove reliance on url pathname in ScrollToTop (#393)
Given that pages are loaded asynchronously, it doesn't make sense for the
window to scroll to the the top once the pathname changes since it doesn't necessarily mean the next/previous page is fully loaded. It's better to rely on a key (scrollKey) and only change the latter when the page is successfuly loaded.