Bug 1536435 - support linked identities
Now that we link identities, the format of the `identities` property has
changed, invalidating some assumptions.  Happily those assumptions were
represented with assertions so we got errors instead of undefined
behavior!

In particular, an account with a github identity now has an identity
like:

```
    {
      "provider": "github",
      "connection": "github",
      "profileData": {
        "userId": 1234,
        "nickname": "developer1234"
      }
    },
```

and an account with a firefoxaccounts sign-in has

```
    {
      "profileData": {
        "email": "foo@bar.com",
        "fxa_sub": "97ff4ba365255c2a92d362ea017baa15",
        ...
      },
      "provider": "oauth2",
      "user_id": "firefoxaccounts|97ff4ba365255c2a92d362ea017baa15",
      "connection": "firefoxaccounts",
      "isSocial": true
    }
```