Bug 1538961 - support acking/nacking messages
This required a major refactor of the AsyncIterator implementation,
including some tests to ensure its functionality.  With this change, the
iterator's pushValue now returns a promise that is only resolved when
the message is consumed.  The result is that amqplib can queue up a
bunch of messages for a slow HTTP client, and those will not be ack'd
until the apollo libraries have deemed the message delivered to the
browser (although who knows how accurate that is..)