Bug 1541741 - add scopesSatisfying
Based on benchmarking, for "small" checks, this function is slightly
faster than `satisfiesExpression`.  For "large" checks, it's about 60%
slower, and most of that is just because it cannot "short circuit" AnyOf
once it has found one satisfied subexpression.

Small checks correspond to most API calls: a scope expression with AllOf
one or two scopes, validated against a large scopeset.  Large checks
correspond to a queue.createTask scope expression, again with a large
scopeset.