Bug 1547000 - don't show diffs for generated files
Note that these files are still shown in a PR, just without the
line-by-line diff -- much like `yarn.lock` is already.

Note, too, that this does not cover files which are *edited* by yarn
generate, such as `package.json` or the READMEs.  Diffs in those files
will appear in PRs as expected.