Rely on `key` and `fetchPolicy` instead of `resetStore`
`resetStore` causes the store to be cleared and all active queries to be refetched. The only problem is that for many of our views, we depend on `fetchMore` to fetch the full list. Refetching active queries will only fetch the first query.

To fix this issue, we should rely on the key prop and the networkPolicy property to refetch active queries. Changing the key prop of a component in React will re-render that component. This will have a similar effect as resetStore.