Bug 1552238 - [BREAKING] refer to workerTypes with workerTypeName
This is of the form `<provisionerId>/<workerType>`, but within the
worker-manager is treated as a single unit.

Note that this is a breaking change, as it changes the indexes for the
Azure tables.  That means all installs will need to delete existing
tables and allow the service to re-create them.  At this point in
deployment of this service, that's not a great harm.

An unfortunate downside of this change is that URLs look like
`/api/worker-manager/v1/worker-type/some-provisioner%7Esome-worker-type`,
with the `/` inside the workerTypeName escaped.  I played around a bit
with the route matching and it seems like `path-to-regexp` (which
Express uses to parse its routes) is completely broken regarding
parameter expressions in the version used by Express 4.  Express 5
should do better, so let's plan to revisit this issue at that time.