Bug 1560649 - generate from local references
This introduce a local HTTP server to satisfy jsonschema2go's desire to
have HTTP URLs, resulting in a much faster run.

Since the URLs are all now to a temporary server on 127.0.0.1, the "See
<url>" bits that jsonschema2go inserts are removed via a regexp.

Rebuilding this content is now as simple as `go generate ./...`, so
`build.sh` is no longer required.