Bug 1560668 - do the whole image build process in one Dockerfile
Multi-stage docker builds are pretty neat.  Turns out, we can do the
whole thing in one Dockerfile, just pre-calculating the TC version and
the entrypoints.

This takes some care to avoid pulling in files that shouldn't be in the
image (like, secrets).  It first filters some obvious things out of the
tarball handed to `docker build`, and then the build process uses `git
clone` to ensure that only files recognized by git are transferred to
the generated image.  The use of multi-stage builds ensures that no
detritus from the build steps is included in the built image.