Background jobs should not create services
The way things are structured, there is only one k8s service per
taskcluster service. Each taskcluster service only has one "web"
deployment, the others are "background" deployments. Previosuly we were
creating duplicate definitions of the same k8s service targeting each
deployment. This means that the one that won was whichever one was
applied last. With this change we only setup services for non-background
deployments, and if there is more than one of those we'll report an
error.