Bug 1572236: Implement TASKCLUSTER_WORKER_LOCATION env variable
The TWC environment variable aims to provide the task some context of
the environment where it is running. It is a json string containing
fields with information about the cloud provider.

The `cloud` key reports which cloud provider it is running.

What is actually given to the worker is a configutation called
`workerLocation` and the worker is responsible to pass as a environment
variable to the tasks.

Ref: https://github.com/taskcluster/taskcluster-rfcs/pull/148/files