Bug 1302596: Retry image download when it fails
We already have retries for when the download fails, but not when the
downloaded image is corrupted.

We add a retry for the general `ensureImage()` function, for edge cases
like above.

As we are not retrying because of possible network temporary outage, we
configure the waiting time to be a small value.