Let renovate run more often and make more PRs
We have a lot of un-updated dependencies.  My guess is that we're
getting two Node and two Golang PR's per week (running once per week
with prHourlyLimit=2), and have been falling behind quickly.  So this
opens things up quite a bit.  If my hunch is correct, and once we get
up-to-date, we can scale this back to run daily so that PRs aren't
popping up mid-day.