Bug 1405711: Fix race condition between worker-shutdown and task running
If we receive a 'graceful-shutdown' message from the worker runner but
we have a task that's claimed but not yet in the list of running tasks,
the task is never aborted and finishes with 'failed' status.

We now check if we received a shutdown message after we add the task to
the list of running tasks and abort it if so. This way there is no more
the condition that happened between task claim and task run.