[regression] Fix deadlock in the container volume test
In the commit 84b1f20ffc we started to wait the tasks to finish before
claiming more work because we needed to check if we received a SIGTERM
signal and exit. This caused a serialization between `runTaskset` and
`getTasks` methods.

In the commit fc1c3a8 we made the two test tasks dependent of each the
other, but `getTasks` might pick only one task for execution and the
other in the next iteration. But the next call to `getTasks` is only
performed after `runTaskset` finishes, and the task will only finish
when the second one start running, causing the deadlock.

We no longer wait for `runTaskset` to finish and call `shouldExit` as a
continuation of the `Promise` returned by `runTaskset`. This wait
`runTaskset` and `getTasks` can run concurrently, eliminating the
deadlock.