Bug 1582376 - use the server default frame_max
amqplib bizarrely defaults to 4k AMQP frames.  The header for each
transmitted message must fit in a single frame, and the header contains
the CC routes, which TC sets based on `task.routes`.

The default frame_max for RabbitAMQP is a much more manageable 128k.

Surprising nobody who's worked with amqplib, when a transmitted message
overflows the frame, the error is delivered via an `error` event on the
connection, but the `chan.publish` callback only gets a sanitary
"channel closed" error.  So we have no way to distinguish this
circumstance from any other.  Per the AMQP protocol, an oversize frame
is a connection-killing event, so this will terminate the connection for
all other users as well -- and does so repeatedly as the publish call is
retried.