Bug 1582376 - refactor publisher retrying
This refactors the retry logic a little bit so that we never try to call
`chan.publish` after the deadline has passed.  In the old way, we would
check the dealdine, then wait for a channel to be available (which can
take several seconds during frequent reconnects), then publish without
checking whether the deadline had passed in the interim.