Add top pagination for table in ConnectionDataTable
Added a top pagination for table in ConnectionDataTable so that
ConnectionDataTable has both top and bottom table paginations by default.
This improves the user experience of paginated tables in that now
users can use the top pagination buttons instead of having to
scroll all the way down to know whether the table is paginated or not.
Also created boolean prop withoutTopPagination to enable opting out
in displaying a pagination at the top.