[UI] Improve ux of a paginated table (ConnectionDataTable) (#1680)
* Create renderTablePagination method in ConnectionDataTable

Separated table pagination component in ConnectionDataTable into its
own render method called renderTablePagination in order to reuse it
for creating table paginations. This enables more easier implementation
of adding both top and bottom table pagination for tables.

* Add top pagination for table in ConnectionDataTable

Added a top pagination for table in ConnectionDataTable so that
ConnectionDataTable has both top and bottom table paginations by default.
This improves the user experience of paginated tables in that now
users can use the top pagination buttons instead of having to
scroll all the way down to know whether the table is paginated or not.
Also created boolean prop withoutTopPagination to enable opting out
in displaying a pagination at the top.

* Change TaskRunsCard to not display pagination on top

Set withoutTopPagination prop value to true in order make sure top
pagination is not displayed in TaskRunsCard component. This results in
displaying small tables with only the bottom pagination.

* Add minor styling changes

* Add changelog