[UI] Fix view title for worker pool view error page (#1694)
* Add disableTitleFormatting prop to Dashboard component

Added disableTitleFormatting prop to Dashboard component in order to
display title as given as is without going through any formatting. At
default this is set to false so that titles are formatted in uppercase.

* Change title format for WMViewErrors

Changed disableTitleFormatting prop value to true in Dashboard component
for WMViewErrors in order to preserve the worker pool ID and not format
it in uppercase. Also uppercased literal part of title and added double quotes
around the worker pool ID to improve the readability of the view title.

* Add changelog referencing issue 1682

Added changelog referencing issue 1682 in order to record silent
changes. This helps the pull request 1694 pass test check for changelog.