Gracefully handle errors when loading tasks
When there is multiple requests to the same endpoint, we batch and use caching to improve performance. Right now if an endpoint throws an error (e.g., fetching a decision task that doesn't exist), then we don't wait until that all of these requests have settled.

This commit makes sure we gracefully handle errors so that all requests
get the chance to resolve.