Create Provisioners Details Table Component
To condense the "provsioners" page, a new provisioner component was
created to replace the old one.
The new provisioner component uses tables to display the provisioner
details instead of cards as suggested in issue #2096.

Currently, the component renders a table with all the provisioners and displays
the id, stability, last available date, and expiration date for each one.
The component also currently sorts by clicking the header.
Actions and description will be displayed in a drawer.

What's left for this component:
  - Drawer still needs to be displayed and handled
  - Actions need to be displayed and handled