refactor Database class, tests
This makes the Database constructor a bit more "private", in that it
expects a set of URLs for each of the "modes" it should support.  In an
upgrade situation, those are 'admin' and 'write', which is a nice
safeguard against using functions in the wrong places.  And it also
fixes the weirdness of supplying a dummy serviceName when performing
upgrades.

The change to the library API is minimal.