Bug 1608185 - proto.WaitUntilInitialized()
This introduces a way to wait for the capabilities negotation to
complete on a protocol, and uses this by default to wait before
verifying `proto.Capable(..)`.  This should help in cases where a
protocol capability is needed early in the worker startup process, such
as reporting a startup error.  It eliminates race conditions that might
otherwise occur in that context.