Bug 1584208 - rewrite client READMEs
With this change, all of the READMEs have the same format and document
their features similarly.  They also have copious reference to the
Taskcluster documentation.  This does drop the inclusion of every API
method in the README for the Python client, as this information is more
accessible in the reference documentation anyway.  Like the go client,
the information remains in the API documentation.