Move taskcluster-go-base into this repo.
This repo depends back onto the Go client, causing version issues when
e.g., v25 of the monorepo refers to a version of tc-base-go that refers
to v24 of the monorepo.

This commit puts the files in `internal/`.  While it might be nice to
have something like `libraries/go` to parallel the JS libraries, the
name `internal` is special to Go and its position here is required in
order that all packages in the monorepo can access it.