Rewrite client-go integration tests
One of the existing tests never ran, as it required real (not temporary)
credentials.  It also required scopes that no longer exist, and involved
a complicated and fragile process of setting up a task.  The other rand
without authentication but accessed an index path that may not exist
(garbage).

These are replaced with an unauthenticated test that will succeed
regardless of the data in the deployment it is run against, and an
authenticated test that uses the auth.TestAuthenticate method developed
for this purpose.