Bug 1616900 - wait for proto initialization in tests
This avoids a subtle race condition where `runnerProto` is not
considered initialized until some time *after* the `hello` message is
received, as the function handling that message must run.

`go test -race` missed this because the use of a Mutex avoids an actual
data race.  The bug here is at a higher level than that.