Bug 1600071: Fix overprovisioning for workers that are slow to boot up
For workers that are booting, worker-manager considers them as "ready"
and possibly running tasks, and it ignores the fact that they will
eventually claim the current pending tasks. This is not a problem if the
instances of these workers boot relatively quickly.

If the instances take a long time to get ready (like baremetal
instances), worker-manager will keep spawning more instances while there
are pending tasks and until we reach the maximum capacity, causing an
overprovisioning problem.

We fix this by subtracting the number of instances in the REQUESTED
state from the total number to spawn.