Bug 1617685 - do not block queue startup on EC2 Region loading
Amazingly, AWS has on several occasions failed to successfully host
their own list of IP address blocks.  With this change, the resolver
tries to load the list, but fails "softly": startup is allowed to
continue, and the resolver will just return "null" for all IPs until the
load is successful.