Grant scope to the worker to terminate itself
We are having issues with the shutdown behavior in AWS instances.
When the worker is done, it calls `shutdown -h now` to stop the
instances and then the `workerManager` can terminate it later. The
problem is that sometimes the shutdown process stalls inside the
instance, hanging the worker forever.

Granting the scope to call `registerWorker` allows the worker to request
`workerManager` to terminate the worker, avoiding the shutdown issue.