Implement "shutdown" message
The "shutdown" message is implemented for AWS, Azure and GCP providers.
When received, worker-runner will call the worker-worker manager
RemoveWorker API. If it fails, it will fall back to the system
shutdown.

For Linux, the "system shutdown" is the execution of the command
"shutdown -h now". For Windows, it will call the Win32 API
InitiateSystemShutdownEx.

The static and standalone providers don't implement the "shutdown"
message.