Bug 1616931 - replace StdioTransport with PipeTransport
StdioTransport implemented io.Reader and io.Writer, but this had the
disadvantage that it was not possible to block waiting to send a
message.  This is problematic when, for example, sending a log message
before exiting a process: in most cases, the process will exit before
the message is sent.

In practice, the API exposed by StdioTransport wasn't that useful,
anyway.  In most cases, we have a reader/writer pair that we want to run
a protocol on, and that required setting up io.Copy operations that run
forever.

PipeTransport's API fixes both: its Send method blocks until the message
is sent, and its top-level API takes a reader and a writer to run on.

This also adds a `protocol/testing` package with utilities for testing
things involving a protocol, and this removes the need for
FakeTransport and simplifies some tests.