Encode string and composite keys and fix handler
Running the importer will import partition and row keys that were
encoded. Because of that, we'll need to keep encoding the values so that
we know how to load or modify an entity using the partition/row keys.

The `this` reference inside the handler was set to global rather than
the subclass. This commit fixes the issue.