Disable tasks on pools with insufficient capacity, linux-arm to tier 2
The worker pools behind these tasks are single workers or otherwise
unable to keep up with demand.  So, rather than merging PRs with yellow
dots or eventually red X's when the tasks time out, let's disable them.

Of these platforms, Windows 10 is a tier-1 platform, but that platform's
testing is covered by the Windows 2012 worker pool.  And linux-arm is a
tier-1 platform, but in the larger Mozilla scope is tier-2 at this
point.