Add opt-out for Matrix notifications being "notices" rather than messages
Riot appears not to trigger a desktop notification when a notice contains my username and a "watched keyword". This is configurable in Riot’s settings, but sending a normal message instead would make it easier for people to be notified of failures they are interested in.

This is a boolean because Matrix’s other message types do not seem relevant here: emote, image, file, audio, location, video.

This commit is a first pass at adding this, however it is done entirely "blind" since my laptop does not have the required Node version, so I can’t easily run tests or regenerate generated files. I’d appreciate someone taking over.