Add a SiteSpecific component and use it in a few places
The component has a distinctive color that will help it stand out as
site-specific.  The text of the component should also suggest this.

The component reads values from deployment config.  If the value isn't
set, then the component does not appear.

I addressed a few instances of root URL in the docs, plus the github app
since it came up in [Bug
1629623](https://bugzilla.mozilla.org/show_bug.cgi?id=1629623).