Refactor fakes for AWS provider
This also factors out some of the provisioning tests into a helper
function that reduces the repetitive setup.

The test for tags was broken (it used the spread operator in the
arguments to `assert.deepEqual` and because the first two elements of
the array it spread were equal, the assert passed, without actually
comparing to the set of tags given).