Rewrite tests for Azure provider
This rewrites the tests to use a more natural set of fakes, and to have
more tests each testing fewer things.  This also adds schema testing for
creation of Azure resources, which will catch any mistakes in specifying
those resources in the future.  Adding tests for new functionality
should be easier, as should modifying tests for changes to
functionality.

A number of minor bugs in the provider are identified with TODO notes in
the tests and schema.s