Prevent onConnected from occuring before a message
This avoids a race condition where a consumer's onConnected callback
could be called *after* the first message is delivered.

This commit also fixes `consumer_test.js` to only start publishing
messages once connected.  This avoids the risk that the messages would
all be published before the consumer began; since this is an ephemeral
consumer, that would mean it receives no messages.