[azure provider] change worker state atomically
Instead of reading the worker's state, doing a bunch of async stuff,
then writing it back to the worker table, this performs state updates
atomically using `worker.update`, and checks that the state always moves
"forward" in the REQUESTED -> RUNNING -> STOPPING -> STOPPED sequence.
This avoids race conditions between, for example, a call to
`registerWorker` (which would set state=RUNNING) and a call to
`removeWorker` (which would set state=STOPPING).

The `XXX note` comment was not true: the provider was incorrectly
setting a worker's state to RUNNING before registerWorker.