Rewrite browser-side auth to fetch credentials on-demand
The previous solution was an attempt to renew credentials on a schedule,
synchronized across tabs.  It had some pretty bad race conditions, one
of which was between page-load and renewal when the existing creds had
expired.

This switches to a simpler model, where credentials are fetched if
necessary (if they're expired) before making a GraphQL query.