Load root certificates from Node.js
tls.rootCertificates contains the root certificates bundled in Node.js,
and sourced from Mozilla's CA store. Load the ones that node-forge can
read into the trusted certificate authority store. Continue loading the
intermediate certs, but check that the issuer root CA is in the store.

Add a test that the Azure root certificates are in the store. Two of
the certs are bundled with Node.js but can't be loaded by node-forge:

* Microsoft RSA Root Certificate Authority 2017
* Microsoft ECC Root Certificate Authority 2017

node-forge supports RSA, so it is unexpected that the first one does
not load. node-forge doesn't support ECC, so the failure with the
second is expected.

Add a helper method dnToString, that converts a distinguished name to
OpenSSL text format (like "/CN=common name").