Download unknown intermediate signing certificates
In the Azure worker manager, when a PKCS#7 signing certificate is
issued by an unknown intermediate certificate, use the
AuthorityAccessInfo extension to find the CA Issue URL, download the
certificate, and verify it was issued by a known root certificate.
A new monitor message, registration-new-intermediate-certificate, has
the details of the certificate. The downloaded certificate is added to
the trusted certificate store for future requests.

Add helper methods specific to Azure and node-forge:

* getCertFingerprint: Get the OpenSSL fingerprint of a certificate
* getAuthorityAccessInfo: Parse AAI into plain object format