Verify attested document signing cert's subject
In the Azure public cloud, the embedded signing certificate in the
PKCS #7 attested data message should have a subject common name (CN)
of "metadata.azure.com", or a subdomain ending in
".metadata.azure.com". This is checked after validating the message
signature.

For tests, a new fixture is added, with a valid PKCS #7 message with
the expected data but signed by a certificate with the CN
metadata.azure.org, not .com. This certificate is issued by a
self-signed certificate, so it would fail the later validation chain as
well.