Add `get_queue_artifacts_paginated` with index pagination
This is the better method of paginating results in a DB, in place of
limit/offset.  It is stable to concurrent insertion and removal of items
from the table, but more importantly it is much more performant.  Where
limit/offset requires starting at the beginning of the table and reading
and discarding _offset_ rows, this query can use the table's primary key
index to find the next row immediately.

I began to add reverse functionality here, but did not find any simple
way to do so in plpgsql without creating queries from strings (`query
execute`), and even doing that is complicated.