Generate signed URLs on the browser at the last minute
Previously, web-server generated signed URLs for artifacts as well as
determining if they're public and if they are a logfile.  With the
advent of non-public deployments, this doesn't work very well anymore.

Instead, we now handle artifacts by taskId / runId / name, and only
generate URLs for them at the last minute.  For Links, this includes
when the link is rendered (in case people copy/paste the URL) and when
the link is clicked.