Fix App's error boundary
Previously, App sent the error to Main as a prop, and Main just
re-rendered the failed tree and tried to layer an error box on top of
it.  Obviously this just threw the same error, and a loop ensued.

With this change, render errors are replaced with a blank page
containing the error message.