Correct online migration functions in versions 59, 60
This fixes the functions to look at the `task_id` property of the
`state`.  Without this change, they look for a property named after the
current task_id, and as that property does not exist, the result is NULL
and the function scans from the beginning of the `tasks` table.