Move schema validity checks from metaschema to validate
Recursive refs are difficult anyway, and the metaschema "cheated" in
ways that Ajv-7.x's strict mode prohibited.  tc-lib-references does lots
of other validation of schemas, so it makes sense to include the checks
for forgotten attributes there as well.

The docker-worker schema was not compliant, and is fixed here.  I
suspect the original authors intended `additionalProperties: false`, but
I also suspect that there are tools out there generating tasks with
additional properties, so it has been left at the default, `true`.

The Kubernetes values schema had a typo, which the validation now
catches.