[azure-provider] Generate names for vm disks
..to allow their deletion during deprovisioning

This also refactors the worker scan such that provisionResources is
called repeatedly until it is finished, even after the VM is registered
as present.  This ensures that `provisionResources` has a chance to see
the registered VM and gather its `id`, even if that occurs after the
worker has started and called `registerWorker`.