Pin Cargo.lock for rust-client
This will prevent "time-based" problems, where some dependency is
upgraded and suddenly CI is broken.  That comes at the expense of not
automatically adopting updates, and of users of the library potentially
using different or newer versions of its dependencies.  But in practice
neither of these are likely to be problematic.