Upgrade httptest and fix tests
The newer version of httptest use the IPv6 localhost address, `::1`.
However, because IPv6 addresses contain `:`, they are written
`[addr]:port` when combined with a port, such as in a URL.  So, this
emulates the same logic for determining the host string in the tests.

In production logic, we would simply trust the `Host` header, so this is
a test-only problem.